quick refactor of files
