--- .deli
@@ -1,8 +1,0 @@
-[DEFAULT]
-DELI_DATA_DIR = ~/.deli/deli_data
-
-BB_MASK = ###
-BB_NULL = NUL
-
-MAX_INDEX_RISK_DIST_THRESHOLD = 3
-MAX_LIBRARY_RISK_DIST_THRESHOLD = 4
