+++ src/deli/base/__init__.py
@@ -1,10 +1,12 @@
 """base __init__"""
 
-from .selection import BaseSelection
-from .settings import BaseSettings
+from .selection import BaseSelection, DecodingExperiment
+from .settings import BaseSettings, DecodingSettings
 
 
 __all__ = [
     "BaseSelection",
     "BaseSettings",
+    "DecodingExperiment",
+    "DecodingSettings",
 ]
