--- src/deli/logging/__init__.py
@@ -1,1 +1,0 @@
-"""init for logging module"""
